requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2
PyJWT==2.10.1
python-decouple==3.8
